Feature/gestion autorisations

+++ time_manager/lib/time_manager_web/controllers/user_controller.ex
@@ -3,6 +3,7 @@ defmodule TimeManagerWeb.UserController do
 
   alias TimeManager.Accounts
   alias TimeManager.Accounts.User
+  import TimeManagerWeb.Authorization
 
   action_fallback TimeManagerWeb.FallbackController
 
@@ -32,8 +33,14 @@ defmodule TimeManagerWeb.UserController do
   end
 
   def index(conn, _params) do
-    users = Accounts.list_users()
-    render(conn, :index, users: users)
+    # if verify_role_super_manager(conn, "super_manager") || verify_role_manager(conn, "manager") do
+      users = Accounts.list_users()
+      render(conn, :index, users: users)
+    # else
+    #   conn
+    #   |> put_status(:unauthorized)
+    #   |> json(%{error: gettext("unauthorized")})
+    # end
   end
 
   def create(conn, %{"user" => user_params}) do
@@ -47,38 +54,63 @@ defmodule TimeManagerWeb.UserController do
   end
 
   def show(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-    render(conn, :show_with_teams, user: user)
+    IO.inspect(verify_role_manager(conn, "manager"))
+    if verify_role_super_manager(conn, "super_manager") || (verify_role_manager(conn, "manager") && is_same_team(conn, id)) || verify_user_id(conn, id) do
+      user = Accounts.get_user!(id)
+      render(conn, :show, user: user)
+    else
+      conn
+      |> put_status(:unauthorized)
+      |> json(%{error: gettext("unauthorized")})
+    end
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Accounts.get_user!(id)
-    if Map.has_key?(user_params, "role") && ! Enum.member?(@roles, user_params["role"]) do
-      raise "role is not correct, please select a correct value"
-    end
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
-      render(conn, :show, user: user)
+    if verify_role_super_manager(conn, "super_manager") || (verify_role_manager(conn, "manager") && is_same_team(conn, id))|| verify_user_id(conn, id) do
+      user = Accounts.get_user!(id)
+      if Map.has_key?(user_params, "role") && ! Enum.member?(@roles, user_params["role"]) do
+        raise "role is not correct, please select a correct value"
+      end
+      with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+        render(conn, :show, user: user)
+      end
+    else
+      conn
+      |> put_status(:unauthorized)
+      |> json(%{error: gettext("unauthorized")})
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
+    if verify_role_super_manager(conn, "super_manager") || (verify_role_manager(conn, "manager") && is_same_team(conn, id))|| verify_user_id(conn, id) do
+      user = Accounts.get_user!(id)
 
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      conn
+      |> put_status(:unauthorized)
+      |> json(%{error: gettext("unauthorized")})
     end
   end
 
 #  get the (+ or -) time credit of the user, in minutes.
   def get_time_credit(conn, %{"id" => id}) do
-    minutes_per_day = 7 * 60
+    if verify_role_super_manager(conn, "super_manager") || (verify_role_manager(conn, "manager") && is_same_team(conn, id))|| verify_user_id(conn, id) do
+      minutes_per_day = 7 * 60
 
-    working_times = TimeManager.WorkingTimes.list_working_times()
-    user_wts = Enum.filter(working_times, fn working_time -> working_time.fk_user == id end)
-    credit_time = Enum.reduce(user_wts, 0, fn working_time, acc ->
-      daily_minutes_working = NaiveDateTime.diff(working_time.end, working_time.start) /60
-      acc + (daily_minutes_working - minutes_per_day)
-    end)
-    render(conn, :show_time_credit, time_credit: credit_time)
+      working_times = TimeManager.WorkingTimes.list_working_times()
+      user_wts = Enum.filter(working_times, fn working_time -> working_time.fk_user == id end)
+      credit_time = Enum.reduce(user_wts, 0, fn working_time, acc ->
+        daily_minutes_working = NaiveDateTime.diff(working_time.end, working_time.start) /60
+        acc + (daily_minutes_working - minutes_per_day)
+      end)
+      render(conn, :show_time_credit, time_credit: credit_time)
+    else
+      conn
+      |> put_status(:unauthorized)
+      |> json(%{error: gettext("unauthorized")})
+    end
   end
 end
